+++ src/features/consent/index.js
@@ -1,0 +1,3 @@
+export { ConsentProvider } from "./context/ConsentContext";
+export { useConsent } from "./hooks/useConsent";
+export { default as ConsentModal } from "./components/ConsentModal";